Added old code from the 2018 MBI workshop
These files implement a basic version of the Couzin model for a group of
agents moving together. The agents try to avoid each other if too close,
align with each other for intermediate distances and move towards each
other if very far apart. Some of the agents are "Leaders" which know of
target destinations such as "Watering holes". Obstacles are modelled as
stationary agents which can only be avoided and have no update rules for
themselves. If the agents don't reach their destination in a given
amount of time they randomly pick a different destination and try to
head towards it.