Added parameter sweep file
This file is an alternative to the analysis file Stefan made. Enter a list of parameters and a number of repetitions for the loop to move through. The results are saved in a folder named Data, within a subfolder corresponding to the parameter set number. Each repetition is saved as a separate .mat file.