Visualizing calovi and couzin models
Short script to visualize the results from calovi and couzin model. You can select a relative agent orientation ('Where is the other agent placed relative to this one') and an absolute agent orientation ('where is the other agent looking') and this visualizer will tell you what the calovi and couzin models say the desired orientation will be.